First pass implementation of the "Tool Meister"
The goal of the "Tool Meister" is to encapsulate the starting and
stopping of tools into a wrapper daemon which can be started once on a
node for the duration of a benchmark script.  Instead of the start/stop
tools scripts using SSH to start/stop tools on local or remote hosts, a
Redis Server is use to communicate with the all the started Tool
Meisters which execute the start/stop operations via messages they
receive vis Redis's publish/subscribe pattern.

The Redis server location is passed as a set of parameters (host & port)
to the tool meister, along with the name of a "key" in the Redis server
which contains the tool meister's initial operating instructions for the
given benchmark script's execution:

  * What Redis pub/sub channel to use
  * What tool group describing the tools to use and their options

The tool meister than runs through a three phase life-cycle until it is
told to terminate: "`start`", "`stop`", and "`send`".  The initial phase
is "`start`", where it waits to be told when to start its tools running
from a published message on the channel it is subscribed. Once it starts
a tool in the background via `screen`, it waits for a "`stop`" message
to invoke its tool's `stop` action.  It then waits for a "`send`"
message to transmit tool data (if any).

We add a simple "data sink" to allow for sending tool data results as
tar balls back to the main controller via HTTP to facilitate the "send"
operation.

** Removal of remote tool registration **

This work IMPLIES that we no longer need to record registered tools
remotely.  We only need to start a tool meister remotely for each host,
passing the initial data it needs at start time via Redis.

Now that pbench-register-tool[-set] supports the ability for a caller to
register a tool [or tool set] for a list of hosts, we can consider
keeping all the tool data locally on the pbench "controller" node where
the pbench-agent's user registers tools.

By doing this, we remove the need to manage a distributed data set
across multiple hosts, allowing for a "late" binding of tools to be run
on a set of hosts.  In other words, the tool registration can be done
without a host being present, with the understanding that it must be
present when a workload is run.

This is particularly powerful for environments like, OpenStack and
OpenShift, where software installation of tools are provided by
container images, VM images (like qcow2), and other automated
installation environments.

This is an invasive change, as knowledge about how tool data is
represented on disk was spread out across different pieces of code. We
have attempted to consolidate that knowledge, future work might be
required to adhere to the DRY principle.

NOTES:

  * The `start` and `stop` messages contain the iteration name (string)
    to use

  * The tool meister invokes the existing tools in `tool-scripts` as
    they operate today without any changes

  * This commit DOES NOT remove pbench-start/stop-tools yet; those
    interfaces are maintained for compatibility initially

  * This work requires further work to eliminate the need to run the
    post-processing step after stopping tools

    - Eliminating the post-processing step is not a requirement, we
      could easily add a "post-processing" phase to the above three
      phases ... but the reason to remove post-processing now is to
      move tool data to more of a pure collection process; this is
      arguable, as we may want some kind of post-processing to
      happen before sending the data back.

TO DO:

Work that is still needed for this commit:

 - [x] Add pbench-tool-meister-start, which will start a Redis server,
       load up all the tool data for the given group into the Redis
       server, start the pbench-tool-data-sink, and start all the tool
       meister's remotely

 - [x] Add pbench-tool-meister-stop, which will stop all the tool
       meisters, stop the data sink, and stop the Redis server

 - [x] Write pbench-tool-miester-client, which will implement the CLI
       method of sending messages via Redis to the tool meisters to
       take a particular action

 - [x] How to get pbench-agent configuration data into python

   - Will likely use the same method we do on the server side

 - [ ] The "`send`" phase is not implemented yet, as we don't have the
       small object store implementation handle that phase yet

 - [ ] Unit tests for this infrastructure

 - [ ] Actually invoke the tool via `screen`

 - [ ] Rewrite pbench-tool-trigger into a python application that
       talks directly to the Redis server to initiate the start, stop,
       send messages

 - [ ] Add support for the tool-meister-start/stop to initiate the
       sysinfo-dump operations