Refactor Tool Meister infra for data capture
The last set of `ssh` commands issued by the pbench-agent that are not
involved in orchestrating the Tool Meisters, outside of those issued by
the `bench-scripts`, are for `hostname -s` data collection on all
the registered remote tool hosts.

The Tool Meisters are now given the responsibility for collecting
information about their environment, including the version of the
pbench agent, and a more comprehensive collection of host name
information, returning that with their startup acknowledgement
payload.

This collected data set is now recorded by the Tool Data Sink, after
restructuring the action flow so that clients send action playloads to
the Tool Data Sink, which then forwards to the Tool Meisters.  This
gives the Tool Data Sink the ability to collect all the Tool Meister
information and write it to disk before acknowledging to the client
that startup is complete.

With this refactoring, we also collapse all the "sysinfo", "init", and
"end" actions into the respective `pbench-tool-meister-start` and
`pbench-tool-meister-stop` interfaces to simplify the CLI behaviors.

To make this possible, we add a new `Client` API used by the CLI
interfaces `pbench-tool-meister-start` and `pbench-tool-meister-stop`.