Refactor Tool Meister infra for data capture
The last set of `ssh` commands issued by the pbench-agent that are not
involved in orchestrating the Tool Meisters, outside of those issued by
the `bench-scripts`, are for `hostname -s` data collection on all
the registered remote tool hosts.

The Tool Meisters are now given the responsibility for collecting
information about their environment, including the version of the
pbench agent, and a more comprehensive collection of host name
information, returning that with their startup acknowledgement
payload.  This change also addresses a previous bug where registered
hosts labels were not being used in the on-disk data capture
directories as they were before.

This collected data set is now recorded by the Tool Data Sink, after
restructuring the action flow so that clients send action playloads to
the Tool Data Sink, which then forwards to the Tool Meisters.  This
gives the Tool Data Sink the ability to collect all the Tool Meister
information and write it to disk before acknowledging to the client
that startup is complete.

We add an entirely new section in the `metadata.log` file which is
per-host/per-tool, recording the tool options, and the output from the
tool install checks (if any).  This section supercedes the previous
tool options in the per-host sections, though we leave those options
there for compatibility for now.

We add graceful handling of bad PUT requests from Tool Meisters to the
Tool Data Sink.

With this refactoring, we also collapse all the "sysinfo", "init", and
"end" actions into the respective `pbench-tool-meister-start` and
`pbench-tool-meister-stop` interfaces to simplify the CLI behaviors.

To make this possible, we add a new `Client` API used by the CLI
interfaces `pbench-tool-meister-start` and `pbench-tool-meister-stop`.

We have also undertaken a major refactoring of `pbench-sysinfo-dump` to
remove the dependency on `base` for environment variables.  As a result
of this work, we drop the stockpile configuration entirely, as it is no
longer required.

Enhance Tool Meister tests to use persistent tools, and enhance the
unit test framework to properly cleans up the Tool Meister test
environment when tests fail.