More code review feedback
We add support for percent-sign ("%") handling in benchmark directory
names and configuration values.

We also had support for recording the label for each registered host.

And we add an entirely new section per-host/per-tool recording the
options, and the output from the install checks (if any).  This section
supercedes the previous tool options in the per-host sections, though
we leave those options there for compatibility for now.