Label support & no `base` in pbench-sysinfo-dump
This commit addresses the need to handle "labels" on registered hosts.

It is also a major refactoring of `pbench-sysinfo-dump` to remove the
dependency on `base` for environment variables.  As a result of this
work, we drop the stockpile configuration entirely, as it is no longer
required.

We also address unit test cleanup of the Tool Meister environment when
tests fail.

And we also add graceful handling of bad PUT requests from Tool Meisters
to the Tool Data Sink.