Remove use of podman by Tool Data Sink
Instead of having the Tool Data Sink leverage `podman` for running a
`prometheus` container, we assume that the `prometheus2` RPM [1] and
PCP RPMs are installed along side the `pbench-agent`.

For non-containerized environments, while the pbench-agent RPM does
not have a direct dependency on the `prometheus2` or PCP RPMs, the
Tool Meister infrastructure will only try to use it if a persistent
tool is registered requiring it.

The "tools" container image builds the `prometheus2`, `node_exporter`,
and PCP RPMs into the image to provide that functionality.

This prevents the pulling images at Tool Meister start-up so that
our callers can avoid unwanted / unexpected behaviors in their
benchmarking environments.

For PCP, removing the use of a `systemd` based PCP container means we
have to now handle creating the N `pmlogger` processes, one for every
`pmcd` executed by Tool Meisters.  We also have to ensure the pmcds are
running before invoking `pmlogger` by using the `pmcd_wait` command.

We provide an updated "pmcd" and "pmlogger" configuration files that
have been tested by hand to show they work even in non-root user
environments.

We also fix the `pmlogger` command to hard-code a 3 second interval for
now.

The Tool Meister PCP tool needs to know the installation directory in
order to find the proper templates to use with `pmcd`.

[1] https://github.com/distributed-system-analysis/prometheus-rpm/pull/1