Remove use of podman by Tool Data Sink
Instead of having the Tool Data Sink leverage `podman` for running a
`prometheus` container, we assume that the `prometheus2` RPM [1] is
installed along side the `pbench-agent`.

For non-containerized environments, while the pbench-agent RPM does
not have a direct dependency on the `prometheus2` RPM, the Tool
Meister infrastructure will only try to use it if a persistent tool
is registered requiring it.

The "tools" container image builds both the `prometheus2` and the
`node_exporter` RPMs into the image to provide that functionality.

This prevents the pulling images at Tool Meister start-up so that
our callers can avoid unwanted / unexpected behaviors in their
benchmarking environments.

[1] https://github.com/distributed-system-analysis/prometheus-rpm/pull/1