Correct the working directory for all tools
We modify the test harness to emit a `<tool>.file` in the current
working directory when the tool runs to help us detect when we don't
have the execution environment correct.  In live testing, lots of
files were being littered in random places.

This led to the discovery that the mock'd `dcgm_prometheus.py` file
was not correct.

We also cleaned up the mocks we had in place to reduce the number of
individual mocks and remove symlinks to the `bench-scripts` tree.

This work also addresses the working directory used by remote Tool
Meisters.