Handle tool shutdown without finite limits
In case the tools we use are not well behaved, the Tool Meisters need
to terminate the sub-processes they create after a short period of
time.

We setup a hard-coded 30 maximum timeout for any tool, persistent or
transient, to shut down, emitting informational messages every 5
seconds while waiting.  After 30 seconds, the sub-process is "killed"
and an error message is emitted.

We also lower the time out in the unit tests since the commands should
not take very long at all.