Refactor PbenchTemplates
Add a new Template DB model that will allow persisting our Elasticsearch
templates to be shared between indexing and the server APIs.

This also refactors the templates.py class which supports indexing, with
the goal of integrating with the new DB table although the base commit
stops short of implementing the integration to allow reviewing the basic
refactoring separately.