Integrate DB into template management
The resolve() method now attempts to look up the root index name in
the templates DB table. If found, it'll check modification timestamps to
determine whether the on disk JSON (currently just resolved and
stat-ed, not read) is more recent. If not, the cached DB information is
used to create a TemplateFile object rather than reading from disk. If
the DB object doesn't exist, it's created; if it exists but is older, it's
updated.